Help to install Salt on Fedora system
Tested with Fedora 42 and so dnf5, enable the SaltStack repository
and install relevant packages.

Clone this repository to /srv/salt if script is called outside it.
Symlink to it if we're on a machine where we want to edit it.

Provide standard minimal configuration for a Salt minion,
so pillar is detected in repository pillar/ subfolder,
and hostname without full domain name is used as id.